requirements.txt
torch==2.6.0
diffusers==0.33.1
Pillow==11.2.1
tqdm==4.67.1
numpy==2.0.2
matplotlib==3.10.0
torchvision==0.21.0